examples/agent-framework-integration/pyproject.toml
[project]
name = "agent-framework-integration"
version = "0.1.0"
description = "an example showcasing integration with microsoft agent framework"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "agent-framework-core>=1.0.0b251114",
    "dotenv>=0.9.9",
    "microsoft-teams-ai",
    "microsoft-teams-apps",
    "microsoft-teams-devtools",
    "microsoft-teams-openai",
]

[tool.uv.sources]
microsoft-teams-apps = { workspace = true }
microsoft-teams-openai = { workspace = true }
microsoft-teams-ai = { workspace = true }
